create slot_machine.py
this game is the basic version of the slot machine where you'll have an default amount of currency 100 using which you can play the game. in each bet the amount you use will be automatically be debited from the default balance and on every spin new balance amout will be displayed which helps you keep track  of balance.
once your balance reach amount: 0 the programme will be ended.

RULES:
if your spin have three same emojis you'll be credited with the amount you won and balance will be updated at that time.